requirements.txt
python-dotenv
pyTelegramBotAPI